requirements.txt
Django==1.10.2
django-bulk-update==1.1.10
django_dotenv==1.4.1
django-extensions==1.6.7
django-jet==1.0.2
django-model-utils==2.6
django-mptt==0.8.6
Scrapy==1.1.3
scrapy-djangoitem==1.1.1